feat: (aflah) Add fullName into User Responses

See merge request enigma-camp/gosling-class/batch-7/tim-challenge/new-group/team-1/backend/be-tekor!106

+++ src/main/java/com/enigma/tekor/service/impl/BundleServiceImpl.java
@@ -134,6 +134,9 @@ public class BundleServiceImpl implements BundleService {
     @Override
     public void delete(UUID id) {
         Bundle bundle = bundleRepository.findById(id).orElseThrow(() -> new NotFoundException("Bundle not found"));
+        if (transactionRepository.existsByBundleId(id)) {
+            throw new ConflictException("Cannot delete bundle that has been purchased");
+        }
         bundleRepository.delete(bundle);
     }
 }
